Corrected some typos. Added a brief description explaining why the Eckart-Young theorem is important.
Added details to how the PCA for any given matrix.

Changed the n_components term in the codes to r_components, so it is in accord with the definition of the Eckward-Young theorem and doesn't create confusion with the n columns of the matrix.

Added an exercise and its solution at the end (I've never used sphinx, so not sure if it's done correctly).